solvers.py modularized
Separated private functions used in solvers.py into a separate _solvers_helpers.py
